[chore] refactor ElasticSwiftCore
