+++ Sources/ElasticSwiftCore/ElasticSwiftCore.swift
@@ -8,6 +8,8 @@
 
 import Foundation
 
+//MARK:- Serializer Protocol
+
 public protocol Serializer {
     
     func decode<T>(data: Data) -> Result<T, DecodingError> where T: Decodable
@@ -16,71 +18,9 @@ public protocol Serializer {
     
 }
 
-public enum MakeBodyError: Error {
-    
-    case noBodyForRequest
-    case wrapped(Error)
-    
-}
+//MARK:- QueryParams enums
 
-public class EncodingError: Error, CustomStringConvertible, CustomDebugStringConvertible {
-    
-    public let value: Any
-    public let error: Error
-    
-    public init(value: Any, error: Error) {
-        self.value = value
-        self.error = error
-    }
-    
-    public var description: String {
-        get {
-            if let val = value as? CustomStringConvertible {
-                return "EncodingError: Unable to encode \(val.description) with Error: \(error.localizedDescription)"
-            }
-            return "EncodingError: Unable to encode \(String(describing: value)) with Error: \(error.localizedDescription)"
-        }
-    }
-    
-    public var debugDescription: String {
-        get {
-            if let val = value as? CustomDebugStringConvertible {
-                return "EncodingError: Unable to encode \(val.debugDescription) with Error: \(error.localizedDescription)"
-            } else {
-                return description
-            }
-        }
-    }
-    
-}
-
-public class DecodingError: Error, CustomStringConvertible, CustomDebugStringConvertible {
-    
-    public let data: Data
-    public let error: Error
-    public let type: Any.Type
-    
-    public init(_ type: Any.Type, data: Data, error: Error) {
-        self.data = data
-        self.error = error
-        self.type = type
-    }
-    
-    public var description: String {
-        get {
-            return "DecodingError: Unable to decode \(String(describing: self.type)) from \(String(describing: self.data)) with Error: \(String(describing: self.error))"
-        }
-    }
-    
-    public var debugDescription: String {
-        get {
-            return "DecodingError: Unable to decode \(self.type) from \(String(reflecting: self.data)) with Error: \(String(reflecting: self.error))"
-        }
-    }
-    
-}
-
-
+/// Enum QueryParams represent all the query params supported by ElasticSearch.
 public enum QueryParams: String {
     case format
     case h
@@ -164,6 +104,7 @@ enum EndPointPath: String {
     case `default` = ""
 }
 
+/// Enum representing URLScheme
 public enum URLScheme: String {
     case http
     case https
