+++ Sources/ElasticSwiftCore/QueryDSL/QueryDSL.swift
@@ -8,9 +8,19 @@
 
 import Foundation
 
+//MARK:- Query Protocol
+
+/// Protocol that all `Query` conforms to
 public protocol Query {
     
     var name: String { get }
     
     func toDic() -> [String: Any]
+}
+
+/// Protocol that all Builder for `Query` conforms to
+public protocol QueryBuilder {
+    
+    var query: Query { get }
+
 }
