+++ Sources/ElasticSwiftCore/Requests/Request.swift
@@ -48,7 +48,9 @@ public class RequestOptions {
     }
 }
 
+//MARK:- RequestBuilder Protocol
 
+/// Protocol to which all `Request` Builders conforms to
 public protocol RequestBuilder {
     
     associatedtype RequestType: Request
